Initial pge module reorganization
