+++ src/opera/pge/__init__.py
@@ -8,6 +8,5 @@ pge
 Contains modules for stand-up and execution of a PGE application.
 """
 
-from .base_pge import PgeExecutor  # noqa: F401
-from .dswx_pge import DSWxExecutor  # noqa: F401
-from .runconfig import RunConfig  # noqa: F401
+from .base.base_pge import PgeExecutor  # noqa: F401
+from .base.runconfig import RunConfig  # noqa: F401
