+++ src/opera/pge/base/schema/catalog_metadata_schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "$id": "https://github.com/nasa/opera-sds-pge/tree/main/src/opera/schema/catalog_metadata_schema.json",
+  "$id": "https://github.com/nasa/opera-sds-pge/tree/main/src/opera/pge/base/schema/catalog_metadata_schema.json",
   "title": "OPERA PGE Catalog Metadata Schema",
   "description": "Schema for validating the catalog metadata output from the OPERA PGEs",
   "definitions": {
